+++ changelog.md
@@ -1,3 +1,9 @@
+### 0.12.0 (Major Release)
+
+#### Adds in a default Lookup List Manager
+The lookup list manager has a `search` and `search_many` methods that allow easy searching across the lookup list itself and any synonyms its members have.
+
+
 ### 0.11.0 (Major Release)
 
 #### Adds options of `today` and `yesterday` in the date picker